Move chain specs to their own modules
